refactor: add folder for classes
